# api/storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"vs-challenge","description":"Test for PHP Developer at VEUS Technology","contact":{"email":"danielsatiro2003@yahoo.com.br"},"license":{"name":"Apache 2.0","url":"http:\/\/www.apache.org\/licenses\/LICENSE-2.0.html"},"version":"1.0.0"},"paths":{"\/api\/v1\/products":{"get":{"tags":["Products"],"summary":"Get products information","description":"Returns products data","operationId":"searchProducts","parameters":[{"name":"q","in":"query","description":"Product name","required":false,"schema":{"type":"string"}},{"name":"filter","in":"query","description":"Search filter","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page number","required":false,"schema":{"type":"integer"}},{"name":"per_page","in":"query","description":"Result number for page","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"successful operation"},"404":{"description":"Product not found"}},"security":[{"passaport":[]}]},"post":{"tags":["Products"],"summary":"Create product","description":"Create product data","operationId":"storeProduct","parameters":[{"name":"id","in":"path","description":"Product id","required":true,"schema":{"type":"integer"}},{"name":"name","in":"query","description":"Product name","required":true,"schema":{"type":"string"}},{"name":"brand","in":"query","description":"Product brand","required":true,"schema":{"type":"string"}},{"name":"price","in":"query","description":"Product price","required":true,"schema":{"type":"float"}},{"name":"amount","in":"path","description":"Product amount","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"successful operation"},"404":{"description":"Product not found"},"412":{"description":"Validation exception"}},"security":[{"passaport":[]}]}},"\/api\/v1\/products\/{id}":{"get":{"tags":["Products"],"summary":"Get product information","description":"Returns product data","operationId":"getProductById","parameters":[{"name":"id","in":"path","description":"Product id","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"successful operation"},"404":{"description":"Product not found"}},"security":[{"passaport":[]}]},"put":{"tags":["Products"],"summary":"Update product information","description":"Update product data","operationId":"updateProduct","parameters":[{"name":"id","in":"path","description":"Product id","required":true,"schema":{"type":"integer"}},{"name":"name","in":"query","description":"Product name","required":true,"schema":{"type":"string"}},{"name":"brand","in":"query","description":"Product brand","required":true,"schema":{"type":"string"}},{"name":"price","in":"query","description":"Product price","required":true,"schema":{"type":"float"}},{"name":"amount","in":"path","description":"Product amount","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"successful operation"},"404":{"description":"Product not found"},"412":{"description":"Validation exception"}},"security":[{"passaport":[]}]},"delete":{"tags":["Products"],"summary":"Delete product information","description":"Delete product data","operationId":"deleteeProduct","responses":{"204":{"description":"successful operation"},"404":{"description":"Product not found"}},"security":[{"passaport":[]}]}},"\/api\/v1\/users\/auth":{"post":{"tags":["User"],"summary":"Get token access","description":"Returns token data","operationId":"auth","parameters":[{"name":"email","in":"query","description":"User email","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","description":"User password","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"successful operation"},"404":{"description":"Product not found"}}}},"\/api\/v1\/users":{"post":{"tags":["User"],"summary":"Create new User","description":"Create user","operationId":"auth","parameters":[{"name":"name","in":"query","description":"User name","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","description":"User email","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","description":"User password","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"successful operation"},"404":{"description":"Product not found"},"412":{"description":"Validation exception"}}}}},"components":{"securitySchemes":{"passaport":{"type":"apiKey","description":"A short description for security scheme","name":"Authorization","in":"header"}}}}